i_sound

+++ src/m_menu.rs
@@ -662,14 +662,14 @@ fn M_DrawSound() {
 			SoundDef.x as usize,
 			SoundDef.y as usize + LINEHEIGHT * (sound_e::sfx_vol as usize + 1),
 			16,
-			snd_SfxVolume,
+			snd_SfxVolume as usize,
 		);
 
 		M_DrawThermo(
 			SoundDef.x as usize,
 			SoundDef.y as usize + LINEHEIGHT * (sound_e::music_vol as usize + 1),
 			16,
-			snd_MusicVolume,
+			snd_MusicVolume as usize,
 		);
 	}
 }
